requirements.txt
streamlit
matplotlib
seaborn
urlextract
wordcloud
pandas
emojiheroku login
